Refactor LueStandardLayout implementation of LueLayout
This commit refactors the implementation of the `LueLayout` trait for the `LueStandardLayout` struct in `layout.rs` file. The changes introduced include the addition of the `host_ref()` and `host_mut()` methods, as well as the `_on_attached_node()`, `_on_detached_node()`, `_on_attached_self()`, and `_on_detached_self()` functions. These changes aim to improve the overall structure and functionality of the `LueLayout` trait implementation.